ref: extract LottieAnimation into its own component
