+++ frontend/src/templates/LottieAnimation/index.ts
@@ -1,0 +1,1 @@
+export * from './LottieAnimation'
